+++ lib/src/models/config/toolbar/buttons/select_header_style_button_configurations.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/widgets.dart' show TextStyle;
+
 import '../../../../widgets/toolbar/base_toolbar.dart';
 
 class QuillToolbarSelectHeaderStyleButtonExtraOptions
@@ -20,9 +22,11 @@ class QuillToolbarSelectHeaderStyleButtonOptions
     super.tooltip,
     this.iconSize,
     this.iconButtonFactor,
+    this.textStyle,
   });
 
   /// By default we will the toolbar axis from [QuillSimpleToolbarConfigurations]
   final double? iconSize;
   final double? iconButtonFactor;
+  final TextStyle? textStyle;
 }