+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+## 9.0.2-dev.1
+* Add configurations for the new dropdown `QuillToolbarSelectHeaderStyleButton`, you can use the orignal one or this
+
 ## 9.0.2-dev
 * **Breaking change** Remove the spacer widget, removed the controller option for each button
 * Add `toolbarRunSpacing` property to the simple toolbar